fix:appClient.ts와 api.ts instance합쳐서 사용 및 네이밍 변경
